Remove db2/util package, move RunAll to db2.Repo as ExecAll
